Move headless.ts under headless directory
