+++ src/main/headless/headless.ts
@@ -1,16 +1,16 @@
 import { ChildProcess, execFileSync } from "child_process";
 import { ipcMain } from "electron";
 import { dirname, basename } from "path";
-import { electronStore, CUSTOM_SERVER, LOCAL_SERVER_URL } from "../config";
+import { electronStore, CUSTOM_SERVER, LOCAL_SERVER_URL } from "../../config";
 import { retry } from "@lifeomic/attempt";
-import { FetchError, HeadlessExitedError } from "../errors";
-import { execute, sleep } from "../utils";
+import { FetchError, HeadlessExitedError } from "../../errors";
+import { execute, sleep } from "../../utils";
 import fetch, { Response } from "electron-fetch";
 import { EventEmitter } from "ws";
 import { BlockMetadata } from "src/interfaces/block-header";
-import { KeyStore } from "./headless/key-store";
-import { Validation } from "./headless/validation";
-import { Apv } from "./headless/apv";
+import { KeyStore } from "./key-store";
+import { Validation } from "./validation";
+import { Apv } from "./apv";
 
 const retryOptions = {
   delay: 100,
